colspan, part 2 https://online.sbis.ru/opendoc.html?guid=6f3117d9-19c1-46fe-8868-92a701d3c587
