+++ Controls/_display/grid/CheckboxCell.ts
@@ -58,19 +58,6 @@ export default class CheckboxCell<T, TOwner extends DataRow<T>> extends Cell<T, 
     shouldDisplayItemActions(): boolean {
         return false;
     }
-
-    getColspan() {
-        return undefined;
-    };
-    getRowspan() {
-        return undefined;
-    };
-    getColspanStyles() {
-        return '';
-    };
-    getRowspanStyles() {
-        return '';
-    };
 }
 
 Object.assign(CheckboxCell.prototype, {
