+++ Controls/_display/grid/HeaderCell.ts
@@ -17,7 +17,7 @@
 */
 
 import { TemplateFunction } from 'UI/Base';
-import {IColspanParams, IHeaderCell} from 'Controls/grid';
+import {IColspanParams, IHeaderCell, IRowspanParams} from 'Controls/grid';
 import HeaderRow from './HeaderRow';
 import { IItemPadding } from '../Collection';
 import Cell, {IOptions as ICellOptions} from './Cell';
@@ -84,13 +84,53 @@ export default class HeaderCell<T> extends Cell<T, HeaderRow<T>> {
         return this._$owner.needMultiSelectColumn() && this._$owner.getHeaderConfig().indexOf(this._$column) === -1;
     }
 
-    _getColspanParams(): Required<IColspanParams> {
-        const params = super._getColspanParams();
-        if (this.isCheckBoxCell()) {
-            params.endColumn--;
-            params.startColumn--;
+    // region Аспект "Объединение колонок"
+    _getColspanParams(): IColspanParams {
+        if (this._$startColumn && this._$endColumn) {
+            return {
+                startColumn: this._$startColumn,
+                endColumn: this._$endColumn
+            };
         }
-        return params;
+        return super._getColspanParams();
+    }
+    // endregion
+
+    // region Аспект "Объединение строк"
+    _getRowspanParams(): Required<IRowspanParams> {
+        const startRow = typeof this._$column.startRow === 'number' ? this._$column.startRow : (this._$owner.getIndex() + 1);
+        let endRow;
+
+        if (typeof this._$column.endRow === 'number') {
+            endRow = this._$column.endRow;
+        } else if (typeof this._$column.rowspan === 'number') {
+            endRow = startRow + this._$column.rowspan;
+        } else {
+            endRow = startRow + 1;
+        }
+
+        return {
+            startRow,
+            endRow,
+            rowspan: endRow - startRow
+        };
+    }
+
+    getRowspan(): number {
+        return this._getRowspanParams().rowspan;
+    }
+
+    getRowspanStyles(): string {
+        if (!this._$owner.isFullGridSupport()) {
+            return '';
+        }
+        const {startRow, endRow} = this._getRowspanParams();
+        return `grid-row: ${startRow} / ${endRow};`;
+    }
+    // endregion
+
+    getWrapperStyles(): string {
+        return super.getWrapperStyles() + ` ${ this.getRowspanStyles() }`;
     }
 
     getWrapperClasses(theme: string, style: string = 'default'): string {
