+++ Controls/_display/grid/mixins/Row.ts
@@ -2,7 +2,6 @@ import { TemplateFunction } from 'UI/Base';
 import { create } from 'Types/di';
 import { IColumn, TColumns, IColspanParams } from 'Controls/_grid/interface/IColumn';
 import { IOptions as IBaseOptions } from '../../CollectionItem';
-import HeaderRow from '../HeaderRow';
 import Collection from '../Collection';
 import Cell, { IOptions as ICellOptions } from '../Cell';
 import { TResultsPosition } from '../ResultsRow';
@@ -10,6 +9,7 @@ import StickyLadderCell from '../StickyLadderCell';
 import CheckboxCell from '../CheckboxCell';
 import prepareColumns from '../../utils/GridColspanUtil';
 import {Model as EntityModel} from 'Types/entity';
+import { THeader } from '../../../_grid/interface/IHeaderCell';
 
 const DEFAULT_GRID_ROW_TEMPLATE = 'Controls/gridNew:ItemTemplate';
 
@@ -123,8 +123,8 @@ export default abstract class Row<T> {
         return this._$owner.getEditingBackgroundStyle();
     }
 
-    getHeader(): HeaderRow<T> {
-        return this._$owner.getHeader();
+    hasHeader(): boolean {
+        return this._$owner.hasHeader();
     }
 
     getResultsPosition(): TResultsPosition {
@@ -225,7 +225,10 @@ export default abstract class Row<T> {
     }
 
     protected _getColspanParams(column: IColumn, columnIndex: number): IColspanParams {
-        return {};
+        const colspanCalculationCallback = this._$owner.getColspanCalculationCallback();
+        if (colspanCalculationCallback) {
+            return colspanCalculationCallback(this.getContents(), column, columnIndex);
+        }
     }
 
     protected _initializeColumns(): void {
@@ -243,11 +246,16 @@ export default abstract class Row<T> {
             for (let columnIndex = 0; columnIndex < this._$columns.length; columnIndex++) {
                 const column = this._$columns[columnIndex];
                 const colspanParams = this._getColspanParams(column, columnIndex);
-                const { startColumn, endColumn, colspan } = colspanParams;
-                if (typeof startColumn === 'number' && typeof endColumn === 'number') {
-                    columnIndex = endColumn - 1;
-                } else if (typeof colspan === 'number') {
-                    columnIndex += colspan - 1;
+                let startColumn, endColumn, colspan;
+                if (colspanParams) {
+                    startColumn = colspanParams.startColumn;
+                    endColumn = colspanParams.endColumn;
+                    colspan = colspanParams.colspan;
+                    if (typeof startColumn === 'number' && typeof endColumn === 'number') {
+                        columnIndex = endColumn - 1;
+                    } else if (typeof colspan === 'number') {
+                        columnIndex += colspan - 1;
+                    }
                 }
                 this._$columnItems.push(factory({
                     column,
@@ -344,6 +352,7 @@ export default abstract class Row<T> {
         return this._$owner.getRowIndex(this);
     }
 
+    abstract getContents(): T;
     abstract getOwner(): Collection<T>;
     abstract getMultiSelectVisibility(): string;
     abstract getTemplate(): TemplateFunction | string;
